feat: add Next.JS to project
