+++ react-components/package.json
@@ -5,19 +5,19 @@
   "homepage": "https://SadJoeBright.github.io/react-gh-pages",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d dist -e react-components",
-    "start": "react-scripts start",
-    "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "format:fix": "npx prettier --write  \"src/**/*.{tsx,html,css}\"",
-    "preview": "vite preview",
     "prepare": "cd .. && husky install react-components/.husky",
     "test": "vitest run --coverage"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.7",
+    "next": "^14.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-loader-spinner": "^5.4.5",
@@ -49,7 +49,6 @@
     "lint-staged": "^15.0.2",
     "prettier": "^3.0.3",
     "typescript": "^5.0.2",
-    "vite": "^4.4.5",
     "vitest": "^0.34.6"
   },
   "lint-staged": {
