--- react-components/src/main.tsx
@@ -1,19 +1,0 @@
-// import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import './index.css';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBounadary';
-import { AppProvider } from './components/context/appContext';
-
-ReactDOM.createRoot(document.getElementById('root')!).render(
-  // <React.StrictMode>
-  <ErrorBoundary>
-    <BrowserRouter>
-      <AppProvider>
-        <App />
-      </AppProvider>
-    </BrowserRouter>
-  </ErrorBoundary>
-  // </React.StrictMode>
-);
