--- react-components/vite.config.ts
@@ -1,15 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  base: './',
-  server: {
-    open: true,
-  },
-  test: {
-    environment: 'jsdom',
-    globals: true,
-  },
-});
